+++ packages/components/form-field/src/index.ts
@@ -1,14 +1,11 @@
 import { FC } from 'react'
 
-import { FormField as Root, FormFieldProps } from './FormField'
-import { FormFieldErrorMessage, FormFieldErrorMessageProps } from './FormFieldErrorMessage'
-import { FormFieldHelperMessage, FormFieldHelperMessageProps } from './FormFieldHelperMessage'
-import { FormFieldLabel, FormFieldLabelProps } from './FormFieldLabel'
-import {
-  FormFieldRequiredIndicator,
-  FormFieldRequiredIndicatorProps,
-} from './FormFieldRequiredIndicator'
-import { FormFieldState, FormFieldStateProps } from './FormFieldState'
+import { FormField as Root, type FormFieldProps } from './FormField'
+import { FormFieldErrorMessage } from './FormFieldErrorMessage'
+import { FormFieldHelperMessage } from './FormFieldHelperMessage'
+import { FormFieldLabel } from './FormFieldLabel'
+import { FormFieldRequiredIndicator } from './FormFieldRequiredIndicator'
+import { FormFieldState } from './FormFieldState'
 
 export { type FormFieldProps } from './FormField'
 export { type FormFieldErrorMessageProps } from './FormFieldErrorMessage'
@@ -17,12 +14,18 @@ export { type FormFieldHelperMessageProps } from './FormFieldHelperMessage'
 export { type FormFieldLabelProps } from './FormFieldLabel'
 export { type FormFieldRequiredIndicatorProps } from './FormFieldRequiredIndicator'
 
+FormFieldLabel.displayName = 'FormField.Label'
+FormFieldState.displayName = 'FormField.State'
+FormFieldErrorMessage.displayName = 'FormField.ErrorMessage'
+FormFieldHelperMessage.displayName = 'FormField.HelperMessage'
+FormFieldRequiredIndicator.displayName = 'FormField.RequiredIndicator'
+
 export const FormField: FC<FormFieldProps> & {
-  Label: FC<FormFieldLabelProps>
-  State: FC<FormFieldStateProps>
-  ErrorMessage: FC<FormFieldErrorMessageProps>
-  HelperMessage: FC<FormFieldHelperMessageProps>
-  RequiredIndicator: FC<FormFieldRequiredIndicatorProps>
+  Label: typeof FormFieldLabel
+  State: typeof FormFieldState
+  ErrorMessage: typeof FormFieldErrorMessage
+  HelperMessage: typeof FormFieldHelperMessage
+  RequiredIndicator: typeof FormFieldRequiredIndicator
 } = Object.assign(Root, {
   Label: FormFieldLabel,
   State: FormFieldState,